Updated plugin to have update queue for the client to read from. Client now gets count of waiting updates and can pull them, by type, one at a time.

+++ Plugin/source/egp-raknet-console/Vector2.cpp
@@ -46,5 +46,5 @@ float Vector2::Distance(Vector2 const &_left, Vector2 const &_right)
 	float y = _right.y - _left.y;
 	y *= y;
 
-	return sqrt(x + y);
+	return (float) sqrt(x + y);
 }
